Dynamic queries

+++ src/main/java/ru/hse/walkplanner/dto/PointDTO.java
@@ -2,6 +2,6 @@ package ru.hse.walkplanner.dto;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
 @Builder
-public record PointDTO(BigDecimal latitude, BigDecimal longitude, String[] tags) {}
+public record PointDTO(double latitude, double longitude, String[] tags) {
+}
